+++ src/ListenerMethods.js
@@ -13,6 +13,7 @@ module.exports = {
      * @param {Action|Store} listenable The listenable we want to search for
      * @returns {Boolean} The result of a recursive search among `this.subscriptions`
      */
+<<<<<<< HEAD
     hasListener: function (listenable) {
         var i = 0,
             listener;
@@ -20,6 +21,17 @@ module.exports = {
             listener = this.subscriptions[i].listenable;
             if (listener === listenable || listener.hasListener && listener.hasListener(listenable)) {
                 return true;
+=======
+    hasListener: function(listenable) {
+        var i = 0, j, listener, listenables;
+        for (;i < (this.subscriptions||[]).length; ++i) {
+            listenables = [].concat(this.subscriptions[i].listenable);
+            for (j = 0; j < listenables.length; j++){
+                listener = listenables[j];
+                if (listener === listenable || listener.hasListener && listener.hasListener(listenable)) {
+                    return true;
+                }
+>>>>>>> spoike/master
             }
         }
         return false;
@@ -73,6 +85,7 @@ module.exports = {
      * @param {Function|String} defaultCallback The callback to register as default handler
      * @returns {Object} A subscription obj where `stop` is an unsub function and `listenable` is the object being listened to
      */
+<<<<<<< HEAD
     listenTo: function (listenable, callback, defaultCallback) {
         var deSub, unsubscriber, subscriptionObj,
             subs = this.subscriptions = this.subscriptions || [];
@@ -82,6 +95,16 @@ module.exports = {
         unsubscriber = function () {
             var index = subs.indexOf(subscriptionObj);
             invariant(index !== -1, 'Tried to remove listen already gone from subscriptions list!');
+=======
+    listenTo: function(listenable, callback, defaultCallback) {
+        var desub, unsubscriber, subscriptionobj, subs = this.subscriptions = this.subscriptions || [];
+        _.throwIf(this.validateListening(listenable));
+        this.fetchInitialState(listenable, defaultCallback);
+        desub = listenable.listen(this[callback]||callback, this);
+        unsubscriber = function() {
+            var index = subs.indexOf(subscriptionobj);
+            _.throwIf(index === -1,'Tried to remove listen already gone from subscriptions list!');
+>>>>>>> spoike/master
             subs.splice(index, 1);
             deSub();
         };
@@ -123,15 +146,20 @@ module.exports = {
     },
 
     /**
-     * Used in `listenTo`. Fetches initial data from a publisher if it has a `getDefaultData` method.
-     * @param {Action|Store} listenable The publisher we want to get default data from
+     * Used in `listenTo`. Fetches initial data from a publisher if it has a `getInitialState` method.
+     * @param {Action|Store} listenable The publisher we want to get initial state from
      * @param {Function|String} defaultCallback The method to receive the data
      */
-    fetchDefaultData: function (listenable, defaultCallback) {
+    fetchInitialState: function (listenable, defaultCallback) {
         defaultCallback = (defaultCallback && this[defaultCallback]) || defaultCallback;
         var me = this;
+<<<<<<< HEAD
         if (_.isFunction(defaultCallback) && _.isFunction(listenable.getDefaultData)) {
             var data = listenable.getDefaultData();
+=======
+        if (_.isFunction(defaultCallback) && _.isFunction(listenable.getInitialState)) {
+            data = listenable.getInitialState();
+>>>>>>> spoike/master
             if (data && _.isFunction(data.then)) {
                 data.then(function () {
                     defaultCallback.apply(me, arguments);
@@ -147,6 +175,7 @@ module.exports = {
      * It will be invoked with the last emission from each listenable.
      * @param {...Publishers} publishers Publishers that should be tracked.
      * @param {Function|String} callback The method to call when all publishers have emitted
+     * @returns {Object} A subscription obj where `stop` is an unsub function and `listenable` is an array of listenables
      */
     joinTrailing: maker("last"),
 
@@ -155,6 +184,7 @@ module.exports = {
      * It will be invoked with the first emission from each listenable.
      * @param {...Publishers} publishers Publishers that should be tracked.
      * @param {Function|String} callback The method to call when all publishers have emitted
+     * @returns {Object} A subscription obj where `stop` is an unsub function and `listenable` is an array of listenables
      */
     joinLeading: maker("first"),
 
@@ -163,6 +193,7 @@ module.exports = {
      * It will be invoked with all emission from each listenable.
      * @param {...Publishers} publishers Publishers that should be tracked.
      * @param {Function|String} callback The method to call when all publishers have emitted
+     * @returns {Object} A subscription obj where `stop` is an unsub function and `listenable` is an array of listenables
      */
     joinConcat: maker("all"),
 
@@ -171,7 +202,7 @@ module.exports = {
      * If a callback triggers twice before that happens, an error is thrown.
      * @param {...Publishers} publishers Publishers that should be tracked.
      * @param {Function|String} callback The method to call when all publishers have emitted
+     * @returns {Object} A subscription obj where `stop` is an unsub function and `listenable` is an array of listenables
      */
     joinStrict: maker("strict")
 };
-