+++ src/ListenerMixin.js
@@ -4,8 +4,13 @@ var merge = require('react/lib/merge'),
 /**
  * A module meant to be consumed as a mixin by a React component. Supplies the methods from
  * `ListenerMethods` mixin and takes care of teardown of subscriptions.
+<<<<<<< HEAD
  *
  * @mixin ListenerMixin
+=======
+ * Note that if you're using the `connect` mixin you don't need this mixin, as connect will
+ * import everything this mixin contains!
+>>>>>>> spoike/master
  */
 module.exports = merge({