+++ src/connect.js
@@ -2,12 +2,22 @@ var _ = require('./utils');
 
 module.exports = function (listenable, key) {
     return {
+        getInitialState: function(){
+            if (!_.isFunction(listenable.getInitialState)) {
+                return {};
+            } else if (key === undefined) {
+                return listenable.getInitialState();
+            } else {
+                return _.object([key],[listenable.getInitialState()]);
+            }
+        },
+
         componentWillMount: function () {
             var me = this,
                 cb = (key === undefined ? this.setState : function (v) {
                     me.setState(_.object([key], [v]));
                 });
-            this.listenTo(listenable, cb, cb);
+            this.listenTo(listenable, cb);
         }
     };
 };