+++ src/createStore.js
@@ -22,8 +22,21 @@ module.exports = function(definition) {
 
     for(var a in Reflux.StoreMethods){
         if (!allowed[a] && (Reflux.PublisherMethods[a] || Reflux.ListenerMethods[a])){
+<<<<<<< HEAD
             throw new Error("Cannot override API method " + a +
                 " in Reflux.StoreMethods. Use another method name or override it on Reflux.PublisherMethods / Reflux.ListenerMethods instead."
+=======
+            throw new Error("Cannot override API method " + a + 
+                " in Reflux.StoreMethods. Use another method name or override it on Reflux.PublisherMethods / Reflux.ListenerMethods instead."
+            );
+        }
+    }
+
+    for(var d in definition){
+        if (!allowed[d] && (Reflux.PublisherMethods[d] || Reflux.ListenerMethods[d])){
+            throw new Error("Cannot override API method " + d + 
+                " in store creation. Use another method name or override it on Reflux.PublisherMethods / Reflux.ListenerMethods instead."
+>>>>>>> spoike/master
             );
         }
     }
@@ -48,6 +61,7 @@ module.exports = function(definition) {
         }
     }
 
+<<<<<<< HEAD
     var context = {};
     mergeInto(context, Reflux.ListenerMethods);
     mergeInto(context, Reflux.PublisherMethods);
@@ -58,6 +72,12 @@ module.exports = function(definition) {
 
     var store = bindMethods(new Store(), definition);
     store.init();
+=======
+    _.extend(Store.prototype, Reflux.ListenerMethods, Reflux.PublisherMethods, Reflux.StoreMethods, definition);
+
+    var store = new Store();
+    bindMethods(store, definition);
+>>>>>>> spoike/master
     Keep.createdStores.push(store);
 
     return store;