+++ src/index.js
@@ -1,7 +1,13 @@
+exports.ActionMethods = require('./ActionMethods');
+
 exports.ListenerMethods = require('./ListenerMethods');
 exports.PublisherMethods = require('./PublisherMethods');
 
 exports.StoreMethods = require('./StoreMethods');
+<<<<<<< HEAD
+=======
+
+>>>>>>> spoike/master
 exports.createAction = require('./createAction');
 exports.createActions = require('./createActions');
 exports.createStore = require('./createStore');
@@ -42,8 +48,16 @@ exports.__keep = require('./Keep');
  * Warn if Function.prototype.bind not available
  */
 if (!Function.prototype.bind) {
+<<<<<<< HEAD
     console.error(
         'Function.prototype.bind not available. ' +
         'ES5 shim required. '
     );
+=======
+  console.error(
+    'Function.prototype.bind not available. ' +
+    'ES5 shim required. ' +
+    'https://github.com/spoike/refluxjs#es5'
+  );
+>>>>>>> spoike/master
 }