+++ test/usingConnectMixin.spec.js
@@ -11,19 +11,36 @@ describe('using the connect(...) mixin',function(){
         assert.equal(connect, Reflux.connect);
     });
 
+    describe("when calling with action",function() {
+        var listenable = {
+                listen: sinon.spy()
+            },
+            context = {setState: sinon.spy()};
+        _.extend(context,connect(listenable));
+
+        it("should pass empty object to state",function(){
+            assert.deepEqual({},context.getInitialState());
+        });
+    });
+
     describe("when calling without key",function(){
-        var defaultdata = "DEFAULTDATA",
+        var initialstate = "DEFAULTDATA",
             listenable = {
                 listen: sinon.spy(),
-                getDefaultData: sinon.stub().returns(defaultdata)
+                getInitialState: sinon.stub().returns(initialstate)
             },
             context = {setState: sinon.spy()},
             result = merge(context, connect(listenable));
 
-        it("should add componentWillMount and WillUnmount",function(){
+        it("should add getInitialState and componentWillMount and WillUnmount",function(){
+            assert.isFunction(context.getInitialState);
             assert.isFunction(context.componentWillMount);
             assert.isFunction(context.componentWillUnmount);
             assert.equal(context.componentWillUnmount,Reflux.ListenerMethods.stopListeningToAll);
+        });
+
+        it("should pass initial state to state",function(){
+            assert.deepEqual(initialstate,context.getInitialState());
         });
 
         result.componentWillMount();
@@ -34,10 +51,6 @@ describe('using the connect(...) mixin',function(){
             assert.equal(context,listenable.listen.firstCall.args[1]);
         });
 
-        it("should pass default data to state",function(){
-            assert.deepEqual([defaultdata],context.setState.firstCall.args);
-        });
-
         it("should store the subscription object correctly",function(){
             assert.equal(listenable,context.subscriptions[0].listenable);
         });
@@ -45,15 +58,19 @@ describe('using the connect(...) mixin',function(){
     });
 
     describe("when calling with key",function(){
-        var defaultdata = "DEFAULTDATA",
+        var initialstate = "DEFAULTDATA",
             triggerdata = "TRIGGERDATA",
             key = "KEY",
             listenable = {
                 listen: sinon.spy(),
-                getDefaultData: sinon.stub().returns(defaultdata)
+                getInitialState: sinon.stub().returns(initialstate)
             },
             context = {setState: sinon.spy()},
             result = merge(context, connect(listenable, key));
+
+        it("should pass initial state to state correctly",function(){
+            assert.deepEqual({KEY:initialstate},context.getInitialState());
+        });
 
         result.componentWillMount();
 
@@ -63,13 +80,9 @@ describe('using the connect(...) mixin',function(){
             assert.equal(context,listenable.listen.firstCall.args[1]);
         });
 
-        it("should pass default data to state correctly",function(){
-            assert.deepEqual([{KEY:defaultdata}],context.setState.firstCall.args);
-        });
-
         it("should send listenable callback which calls setState correctly",function(){
             listenable.listen.firstCall.args[0](triggerdata);
-            assert.deepEqual([_.object([key],[triggerdata])],context.setState.secondCall.args);
+            assert.deepEqual([_.object([key],[triggerdata])],context.setState.firstCall.args);
         });
     });
     describe("when calling with falsy key",function(){
@@ -82,6 +95,17 @@ describe('using the connect(...) mixin',function(){
         it("should send listenable callback which calls setState correctly",function(){
             listenable.listen.firstCall.args[0](triggerdata);
             assert.deepEqual([_.object([key],[triggerdata])],context.setState.firstCall.args);
+        });
+    });
+    describe("together with ListenerMixin in a React component",function(){
+        var store = Reflux.createStore({}),
+            def = {setState:function(){}},
+            fakecomponent = _.extend(def,Reflux.connect(store),Reflux.ListenerMethods);
+        it("should log a warning)",function(){
+            sinon.spy(console,"warn");
+            fakecomponent.componentWillMount();
+            assert(console.warn.calledOnce);
+            console.warn.restore();
         });
     });
 });